+++ api/team.ts
@@ -1,5 +1,9 @@
 import { AxiosError, AxiosInstance } from "axios";
-import { TeamData, TeamRegistrationForm } from "../interfaces/team";
+import {
+  TeamData,
+  TeamDetailData,
+  TeamRegistrationForm,
+} from "../interfaces/team";
 import { ApiError, StandardError } from "./error";
 
 export const createTeam = async (
@@ -27,4 +31,46 @@ export const getTeam = async (axios: AxiosInstance): Promise<TeamData> => {
     });
 };
 
-// export const getTeamDetail
+export const getTeamDetail = async (
+  axios: AxiosInstance,
+  teamId: number
+): Promise<TeamDetailData> => {
+  return axios
+    .get<TeamDetailData>(`/competition/teams/${teamId}/`)
+    .then((response) => {
+      return response.data;
+    })
+    .catch((error: AxiosError) => {
+      throw new ApiError<StandardError>(StandardError.ERROR, error.message);
+    });
+};
+
+export const putTeam = async (
+  axios: AxiosInstance,
+  teamForm: TeamRegistrationForm,
+  teamId: number
+): Promise<TeamData> => {
+  return axios
+    .put<TeamData>(`/competition/teams/${teamId}`, teamForm)
+    .then((response) => {
+      return response.data;
+    })
+    .catch((error: AxiosError) => {
+      throw new ApiError<StandardError>(StandardError.ERROR, error.message);
+    });
+};
+
+export const editTeam = async (
+  axios: AxiosInstance,
+  teamForm: Partial<TeamRegistrationForm>,
+  teamId: number
+): Promise<TeamData> => {
+  return axios
+    .put<TeamData>(`/competition/teams/${teamId}`, teamForm)
+    .then((response) => {
+      return response.data;
+    })
+    .catch((error: AxiosError) => {
+      throw new ApiError<StandardError>(StandardError.ERROR, error.message);
+    });
+};
