--- interfaces/teamMember.ts
@@ -1,8 +1,0 @@
-export interface TeamMember {
-  id: number;
-  fullName: string;
-  email: string;
-  hasAccount: string;
-  isTeamleader: string;
-  createdAt: string;
-}
